requirements.txt
Django==4.2.14
djangorestframework==3.14.0
-djangorestframework-jwt==1.11.0
+djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
redis==5.0.1
Pillow==10.2.0 
drf-yasg==1.21.5

qrcode~=7.4.2